+++ lib/aka/post_message.rb
@@ -1,16 +1,21 @@
 module Post
-  MESSAGE = "## v0.1.10
+  MESSAGE = "
+
+  ## New commands
+  * `aka proj`
+  * `aka proj --save`
+  * `aka proj --load`
+
+  ## v0.1.10
   * Added [aka proj --export] to export alias groups into a project directory
   * Added [aka proj --export ] to generate specific aka file
   * Added [aka proj --import ] to generate specific aka file
   * Added [aka proj] to show all the aliases for current project directory
   * Added [aka groups] to list all groups
-  * Edited test_aka.rb
-  * Added .profile as setup options. Thanks to
-  * Remove some dependencies and set required ruby back to 2.0.0.
-  * Change default list to 20 aliases
-  * Start using aka in your projects :)
+  * Added .profile as setup options. Thanks to fgperianez
 
-  To contributing to this project, send us your pull request ->  https://github.com/ytbryan/aka
+  For bugs/issues, please submit an issue at -> https://github.com/ytbryan/aka/issues
+
+  To contributing to this project, send us your pull request ->  https://github.com/ytbryan/aka/pulls
   "
 end